+++ packages/ses/test/lockdown-options.test.js
@@ -3,13 +3,23 @@ import { repairIntrinsics } from '../src/lockdown.js';
 
 test('repairIntrinsics throws with non-recognized options', t => {
   t.throws(
-    () => repairIntrinsics({ mathTaming: 'unsafe', abc: true }),
+    () => repairIntrinsics({ abc: true }),
     undefined,
     'throws with value true',
   );
   t.throws(
-    () => repairIntrinsics({ mathTaming: 'unsafe', abc: false }),
+    () => repairIntrinsics({ abc: false }),
     undefined,
     'throws with value false',
+  );
+  t.throws(
+    () => repairIntrinsics({ mathTaming: 'unsafe' }),
+    undefined,
+    'throws with deprecated option mathTaming',
+  );
+  t.throws(
+    () => repairIntrinsics({ dateTaming: 'unsafe' }),
+    undefined,
+    'throws with deprecated option dateTaming',
   );
 });